Begin major UI rework
