+++ src/components/SymmetryGroupDisplay.vue
