+++ src/components/TopBar.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <div class="flex flex-row items-center my-2">
+    <h1 class="flex-1 font-bold">Polytwisters</h1>
+    <div class="flex-1 flex flex-row items-center justify-end gap-2">
+      <a
+        target="_blank"
+        href="https://github.com/polytwisters/polytwisters.com/"
+      >
+        source code
+      </a>
+    </div>
+  </div>
+</template>
