+++ vite.config.ts
@@ -7,6 +7,11 @@ import tailwindcss from "@tailwindcss/vite";
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [vue(), tailwindcss(), VitePWA()],
+  resolve: {
+    alias: {
+      "@": resolve(__dirname, "src"),
+    }
+  },
   build: {
     rollupOptions: {
       input: {
